pyproject.toml
[project]
name = "onsides"
version = "3.1.0"
description = "Database of drug adverse events from product labels"
readme = "README.md"
authors = [
    { name = "Nicholas Tatonetti" },
    { name = "Yutaro Tanaka" },
    { name = "Michael Zietz" }
]
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "connectorx>=0.4.2",
    "duckdb>=1.2.2",
    "fastexcel>=0.13.0",
    "httpx>=0.28.1",
    "lxml>=5.3.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "polars>=1.22.0",
    "pyahocorasick>=2.1.0",
    "pyarrow>=19.0.0",
    "pydantic>=2.10.6",
    "pypdf2>=3.0.1",
    "ratelimit>=2.2.1",
    "rich>=13.9.4",
    "scikit-learn>=1.6.1",
    "snakemake>=8.29.3",
    "sqlalchemy[asyncio]>=2.0.38",
    "sqlmodel>=0.0.22",
    "tabula-py>=2.10.0",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
    "transformers>=4.50.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
annotator = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[project.scripts]
build-zip = "onsides.cli:main"
onsides-construct-ref = "onsides.construct_training_data:main"
onsides-train = "onsides.train:main"
onsides-compute-flags = "onsides.compute_annotation_flags:main"
onsides-annotate = "onsides.annotator.app:main"
